Added middleware and pipelines
Middleware is now written which stops things downloading if the url is already in the
database. This connects to the database and searches for a list of urls which is then
checked against. If in it then not downloaded else the response is downloaded.

Pipelines. This class now contains code to open a connection to the database and keeps
it open. It also has a function to get either a list of uni_names or main subject areas
which can be decided up front by the two boolean arguments unis and subs. Code started
to insert an item into the database for the uni_ids and the sub_ids.